+++ core/rewards/contribution_reward_calculator.go
@@ -84,6 +84,8 @@ func calculateRewardsByContributionIndividual(epochSeq, asset, accountID string,
 		partyScores = proRataRewardCalculator(partyContribution, rewardFactors)
 	} else if ds.DistributionStrategy == vega.DistributionStrategy_DISTRIBUTION_STRATEGY_RANK {
 		partyScores = rankingRewardCalculator(partyContribution, ds.RankTable, rewardFactors)
+	} else if ds.DistributionStrategy == vega.DistributionStrategy_DISTRIBUTION_STRATEGY_RANK_LOTTERY {
+		partyScores = rankingLotteryRewardCalculator(partyContribution, ds.RankTable, rewardFactors, timestamp)
 	}
 
 	cap := num.DecimalZero()